+++ packages/components/Button/src/FAB/FAB.tsx
@@ -1,11 +1,10 @@
 /** @jsx withSlots */
 import * as React from 'react';
 import { Pressable } from 'react-native';
-import { fabName, FABType } from './FAB.types';
+import { fabName, FABProps, FABType } from './FAB.types';
 import { TextV1 as Text } from '@fluentui-react-native/text';
 import { compose, UseSlots } from '@fluentui-react-native/framework';
 import { Icon } from '@fluentui-react-native/icon';
-import { ButtonCoreProps } from '../Button.types';
 
 export const FAB = compose<FABType>({
   displayName: fabName,
@@ -14,8 +13,8 @@ export const FAB = compose<FABType>({
     icon: Icon,
     content: Text,
   },
-  useRender: (_userProps: ButtonCoreProps, _useSlots: UseSlots<FABType>) => {
-    return (_final: ButtonCoreProps, ..._children: React.ReactNode[]) => {
+  useRender: (_userProps: FABProps, _useSlots: UseSlots<FABType>) => {
+    return (_final: FABProps, ..._children: React.ReactNode[]) => {
       return null; // Only implemented for mobile endpoints
     };
   },