+++ packages/components/text/src/Variants.ts
@@ -2,6 +2,8 @@ import { Text } from './Text';
 
 // stubbed out for other platforms for which variant alias tokens aren't defined
 export const Caption1 = Text;
+export const Caption1Strong = Text;
+export const Caption2 = Text;
 export const Body1 = Text;
 export const Body1Strong = Text;
 export const Body2 = Text;
@@ -12,5 +14,7 @@ export const Subtitle2 = Text;
 export const Subtitle2Strong = Text;
 export const Title1 = Text;
 export const Title1Strong = Text;
+export const Title2 = Text;
+export const Title3 = Text;
 export const LargeTitle = Text;
 export const Display = Text;