+++ packages/deprecated/theming-ramp/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @uifabricshared/theming-ramp
 
-This log was last generated on Thu, 14 Jul 2022 18:09:50 GMT and should not be manually modified.
+This log was last generated on Wed, 21 Sep 2022 21:00:40 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.2
+
+Wed, 21 Sep 2022 21:00:40 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.19.1
+
+## 0.17.1
+
+Fri, 16 Sep 2022 01:53:06 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.19.0
 
 ## 0.17.0